+++ test/integration_test.sh
@@ -24,6 +24,9 @@ ASCIIGenome="$1 --debug 2 -ni"
 
 set -x
 
+## Can show read pairs
+$ASCIIGenome -x 'readsAsPairs' ../test_data/pairs.sam > /dev/null
+
 ## Init from VCF
 $ASCIIGenome ../test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz -x 'show genome' > /dev/null
 $ASCIIGenome https://raw.githubusercontent.com/dariober/ASCIIGenome/master/test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz -x 'show genome' > /dev/null